+++ tweet_semantic_distance_xls_combination.py
@@ -2,7 +2,7 @@ import json
 import xlrd
 import numpy as np
 import matplotlib.pyplot as plt
-from test_kernel_functions import get_kernel
+# from test_kernel_functions import get_kernel
 
 from src.nlp.SemanticString import SemanticString
 # from src.visualization.visualization import SemanticVisualization
@@ -11,25 +11,25 @@ filename = './data/semantic-distance-database.json'
 READ = 'rb'
 db = json.load(open(filename,READ))
 
-db_name = 'db-text'
-kernel = get_kernel(db_name,normed=True)
+db = {}
+wordfreq = {}
 
-txt_name = './tweet_comparison_data/drink_20.txt'
+txt_name = './tweet_comparison_data/drink_10.txt'
 with open(txt_name) as f:
   strings = [tweet.strip() for tweet in f.readlines()]
 
-lemmaStrings = [SemanticString(string, db, kernel).lemma() for string in strings]
+lemmaStrings = [SemanticString(string, db, wordfreq).lemma() for string in strings]
 
-wb_name = 'tweet_comparison_data/tweet20_comparison_total.xls'
+wb_name = 'tweet_comparison_data/tweet10_comparison_total.xls'
 workbook_input = xlrd.open_workbook(wb_name)
 sheet_input = workbook_input.sheet_by_index(0)
 
-ntweets = {'10_tweets': {'toxic': 1, 'tru': 2, 'mike': 3, 'nick': 4, 'toxicNEW': 5, 'toxicKernel': 6},
-			'20_tweets': {'toxic': 1, 'nick': 2, 'tru': 3, 'mike': 4, 'toxicNEW': 5, 'toxicKernel': 6}}
+ntweets = {'10_tweets': {'toxic': 1, 'tru': 2, 'mike': 3, 'nick': 4, 'toxicNEW': 5, 'toxicKernel': 6, 'toxicEntropy': 9},
+			'20_tweets': {'toxic': 1, 'nick': 2, 'tru': 3, 'mike': 4, 'toxicNEW': 5, 'toxicKernel': 6, 'toxicEntropy': 9}}
 
-raters = ntweets['20_tweets']
+raters = ntweets['10_tweets']
 
-rater = 'toxicKernel'
+rater = 'toxicEntropy'
 
 data = [sheet_input.cell_value(row, raters[rater]) for row in xrange(1, sheet_input.nrows)]
 
@@ -47,7 +47,7 @@ matrix[np.diag_indices(len(lemmaStrings))] = 0
 
 sizes = {'10_tweets': (10,10), '20_tweets': (15,15)}
 
-size = sizes['20_tweets']
+size = sizes['10_tweets']
 fig = plt.figure(figsize=size)
 plt.xticks(range(len(lemmaStrings)+1))
 plt.yticks(range(len(lemmaStrings)+1))
